Post testeo Postman
Cambios Pequeños
Estructura y lógica creada. Testeo con éxito

+++ Semana8/Microservicio-Productos/src/main/java/com/curso/dao/ProductoDao.java
@@ -27,5 +27,5 @@ public interface ProductoDao extends JpaRepository<Producto, Integer> {
 	 * @return
 	 */
 	@Query("SELECT p.precio FROM Producto p WHERE p.codProducto = :codProducto")
-	int obtenerPrecio(@Param("codProducto") int codProducto);
+	double obtenerPrecio(@Param("codProducto") int codProducto);
 }
